Fix incorrect use of InputLoopService in simple ASQ transport sample
